Move analytics_task into celery app
